Add files via upload
At the moment it generates an array of arrays of Tile objects. Tiles have only one attribute right now, it's an int field called tileContent. We can add more.

So it's a grid of tile's with int values. I'm generating them all as 1's right now to represent an empty tile.

The general WASD controls aren't set up yet. The next steps for me are to create a "character" class, it will just be an int "2" for now. Have it "spawn" when the grid is created, and tie WASD to flipping the tiles int's as they are pressed.